+++ block/src/woocommerce-tiered-price-table.js
@@ -1,8 +1,10 @@
-import TieredPriceTableCompatibility from './includes/TieredPriceTableCompatibility';
+import TieredPriceTable from './compatibility/TieredPriceTable';
 
 window.addEventListener( 'DOMContentLoaded', () => {
 	const viewConfig = window.wtsViewObject || {
 		originalTaxDisplay: 'incl',
 	};
-	TieredPriceTableCompatibility.initialize( viewConfig.originalTaxDisplay );
+
+	const tpt = new TieredPriceTable( viewConfig.originalTaxDisplay );
+	tpt.init();
 } );