+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Changelog
 All notable changes to the Wdevs Tax Switch plugin will be documented in this file.
 
+
+## [1.1.3] - 2024-10-15
+### Added
+- Compatibility for [Discount Rules for WooCommerce](https://wordpress.org/plugins/woo-discount-rules/)
+### Updated
+- Migrated filters from only using 'woocommerce_get_price_html' to 'woocommerce_get_price_html' in combination with 'wc_price'. This way, the plugin should be more compatible with other plugins.
+- Compatibility for Tiered Pricing Table for WooCommerce on catalog pages
+
+## [1.1.2] - 2024-10-09
+### Added
+- Compatibility for Measurement Price Calculator for WooCommerce
+
 ## [1.1.1] - 2024-10-05
 ### Added
 - Compatibility for Tiered Pricing Table for WooCommerce Premium
-
 
 ## [1.1.0] - 2024-10-04
 ### Added